refactor: extract duplicated helper functions to shared utils module
- Create new utils.py module in api/routes with shared helper functions
- Move _parse_namespace, _quote_identifier, _build_namespace_path functions
- Move UNIT_SEPARATOR constant to utils module
- Update ui.py to import helpers from utils instead of local definitions
- Update catalog.py to import helpers from utils instead of local definitions
- Eliminates code duplication between ui.py and catalog.py
- All 88 tests pass (catalog + ui)
